[SUS-46] Feat/signup_token_issue_added
