+++ index.ts
@@ -1,6 +1,9 @@
 import { testPostgresConnection } from '#config/postgres'
 import globalExceptionHandler from '#error/globalExceptionHandler'
+import { memberRouter } from '#member/router'
 import { mockRouter } from '#mock/router'
+import { loginRouter } from '#socialLogin/router'
+import cookieParser from 'cookie-parser'
 import cors from 'cors'
 import { configDotenv } from 'dotenv'
 import express from 'express'
@@ -17,6 +20,8 @@ app.use(
 )
 
 app.use(express.json())
+app.use('/member', memberRouter)
+app.use('/login', loginRouter)
 app.use('/mock', mockRouter)
 app.use(globalExceptionHandler)
