+++ src/core/util/nickname/const/modifierList.ts
