+++ src/core/util/nickname/nicknameGenerator/index.ts
