+++ src/socialLogin/dto/frontend/response/loginUserDataResponse.ts
